Add files via upload
This is an airlines database I made in behalf of the subject "Databases 2" for my University. I was asked to create Database in which it contains 3 tables with foreign and primary keys for each one and same data in the tables. The code also has select commands that shows every reservation for every flight and an other one that sorts the customers by the cost of their flights in descending order. The table Flights has an updateable view and the table Reservations has an insertable view.